refactor: clean code

+++ pkg/core/ssl-applicator/acme-dns01/providers/aliyun-esa/internal/lego.go
@@ -36,7 +36,7 @@ type Config struct {
 
 	PropagationTimeout time.Duration
 	PollingInterval    time.Duration
-	TTL                int32
+	TTL                int
 	HTTPTimeout        time.Duration
 }
 
@@ -50,7 +50,7 @@ type DNSProvider struct {
 
 func NewDefaultConfig() *Config {
 	return &Config{
-		TTL:                int32(env.GetOrDefaultInt(EnvTTL, 300)),
+		TTL:                env.GetOrDefaultInt(EnvTTL, 300),
 		PropagationTimeout: env.GetOrDefaultSecond(EnvPropagationTimeout, 2*time.Minute),
 		PollingInterval:    env.GetOrDefaultSecond(EnvPollingInterval, dns01.DefaultPollingInterval),
 		HTTPTimeout:        env.GetOrDefaultSecond(EnvHTTPTimeout, 30*time.Second),
@@ -239,7 +239,7 @@ func (d *DNSProvider) addOrUpdateDNSRecord(siteId int64, effectiveFQDN, value st
 		Data: &aliesa.CreateRecordRequestData{
 			Value: tea.String(value),
 		},
-		Ttl: tea.Int32(d.config.TTL),
+		Ttl: tea.Int32(int32(d.config.TTL)),
 	}
 	_, err = d.client.CreateRecord(request)
 	return err
