+++ pkg/sdk3rd/rainyun/api_ssl_center_get.go
@@ -12,11 +12,11 @@ type SslCenterGetResponse struct {
 	Data *SslDetail `json:"data,omitempty"`
 }
 
-func (c *Client) SslCenterGet(sslId int32) (*SslCenterGetResponse, error) {
+func (c *Client) SslCenterGet(sslId int64) (*SslCenterGetResponse, error) {
 	return c.SslCenterGetWithContext(context.Background(), sslId)
 }
 
-func (c *Client) SslCenterGetWithContext(ctx context.Context, sslId int32) (*SslCenterGetResponse, error) {
+func (c *Client) SslCenterGetWithContext(ctx context.Context, sslId int64) (*SslCenterGetResponse, error) {
 	if sslId == 0 {
 		return nil, fmt.Errorf("sdkerr: unset sslId")
 	}
